# services/production-dashboard/openapi.yaml
openapi: 3.0.0
info:
  title: Production Dashboard API
  description: WebSocket + REST API for real-time production updates in PrintShop OS
  version: 1.0.0
  contact:
    name: PrintShop OS Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: ws://localhost:3000
    description: WebSocket server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Machine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [screen_press, dtg_printer, embroidery_machine, curing_station]
        status:
          type: string
          enum: [idle, running, down, maintenance]
        currentJobId:
          type: string
        utilizationPercent:
          type: number
        lastMaintenance:
          type: string
          format: date-time
        uptime:
          type: number
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [clocked_in, clocked_out, break]
        currentJobId:
          type: string
        clockedInAt:
          type: string
          format: date-time
        productivity:
          type: number
        skillSet:
          type: array
          items:
            type: string
...
    RESTResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        timestamp:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /api/production/resources/machines/{machineId}:
    get:
      tags:
        - Resources
      summary: Get specific machine
      description: Get details of a specific machine
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Machine details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Machine'
        '404':
          description: Machine not found
  /api/production/resources/machines/{machineId}/allocate:
    post:
      tags:
        - Resources
      summary: Allocate machine to order
      description: Allocate a machine to a specific order (supervisor and admin only)
      parameters:
        - name: machineId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
      responses:
        '200':
          description: Machine allocated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Machine'
        '403':
          description: Insufficient permissions
...
  /api/production/resources/employees/{employeeId}:
    get:
      tags:
        - Resources
      summary: Get specific employee
      description: Get details of a specific employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
  /api/production/resources/employees/{employeeId}/assign:
    post:
      tags:
        - Resources
      summary: Assign employee to order
      description: Assign an employee to a specific order (supervisor and admin only)
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: string
      responses:
        '200':
          description: Employee assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/RESTResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Employee'
        '403':
          description: Insufficient permissions
...